+++ src/components/Fields/Fields.js
@@ -2,12 +2,14 @@
 import * as React from 'react';
 import { Field } from 'redux-form';
 import { injectIntl } from 'react-intl';
+import { Col, Row } from 'reactstrap';
 
 import Simple from './Simple';
 import Select from './Select';
 import Rating from './Rating';
 import Geography from './Geography';
 import Varietals from './Varietals';
+import DateTime from './DateTime';
 
 const ConnectedField = (component) => injectIntl((props) => (
     <Field
@@ -54,3 +56,4 @@ export const RatingField = ConnectedRatingField(Rating);
 export const CountryField = ConnectedField(Geography.CountryInput);
 export const RegionField = ConnectedField(Geography.RegionInput);
 export const VarietalsField = ConnectedField(Varietals);
+export const DateTimeField = ConnectedField(DateTime.DateTimeInput);
