+++ src/components/Wines/Form/index.js
@@ -1,0 +1,4 @@
+// @flow
+import Form from './Form';
+
+export default Form;
